remote: move StoreReply to Types.StoreReply
